refactor: consolidate person detail actions into a single dropdown menu
Replace three separate header buttons (Find Duplicates, Edit, Delete) and
the bottom-of-page VCF export with a unified three-dots action menu next
to the Edit button. The new PersonActionsMenu component absorbs all logic
from FindDuplicatesButton, PersonVCardExport, and DeletePersonButton.
